correct warning

+++ demos_cpp/imfiledialog_test/imfiledialog_test.main.cpp
@@ -22,7 +22,7 @@ void setup_texture_loader()
         glGenerateMipmap(GL_TEXTURE_2D);
         glBindTexture(GL_TEXTURE_2D, 0);
 
-        return (void*)tex;
+        return (void*)(size_t)tex;
     };
 
     ifd::FileDialog::Instance().DeleteTexture = [](void* tex) {
